+++ instrumentation-loongsuite/loongsuite-instrumentation-langchain/CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Langchain embedding span support([#157](https://github.com/alibaba/loongsuite-python-agent/pull/157))
+- LangChain embedding span support
+  ([#157](https://github.com/alibaba/loongsuite-python-agent/pull/157))
 - Rerank / document-compressor span support
   ([#149](https://github.com/alibaba/loongsuite-python-agent/pull/149))